fix : process monitoring

+++ skewer.go
@@ -24,16 +24,19 @@ func Listen(opts ...config.Option) error {
 
 	conf := config.Get()
 
+	var ch chan error
+
+	terminal.Start(conf.Verbose)
+
 	go func() {
-		err = monitoring(conf.Args)
+		err = monitoring(conf.Args, conf.IgnoreFiles)
+		ch <- err
 		if err != nil {
 			log.Printf("monitoring error: %+v\n", err)
 		}
 	}()
 
 	//TODO goが存在するかの確認
-	terminal.Start()
-
 	bin := conf.Bin
 
 	//シグナル待受
@@ -48,7 +51,14 @@ func Listen(opts ...config.Option) error {
 	}()
 
 	setStatus(WaitingForRebootStatus)
-	go rebuildMonitor(10)
+	go rebuildMonitor(5, ch)
+
+	//TODO ch を監視
+	go func() {
+		err := <-ch
+		if err != nil {
+		}
+	}()
 
 	return startProxyServer(conf.Port)
 }
@@ -63,13 +73,12 @@ func checkConnection(port int) (chan bool, error) {
 	go func() {
 		var conn net.Conn
 		var err error
-		for range time.Tick(20 * time.Millisecond) {
+		for range time.Tick(100 * time.Millisecond) {
 			conn, err = net.Dial("tcp", fmt.Sprintf(":%d", port))
 			if err == nil {
-				log.Println("ok")
 				break
 			}
-			log.Println("err")
+			log.Println(err)
 		}
 		defer conn.Close()
 		ch <- true
@@ -89,47 +98,49 @@ func cleanup(bin string) {
 	//TODO 付け焼き刃
 	time.Sleep(1 * time.Second)
 
-	err = os.Remove(bin)
-	if err != nil {
-		log.Println(err)
+	if _, err := os.Stat(bin); err == nil {
+		err = os.Remove(bin)
+		if err != nil {
+			log.Println(err)
+		}
 	}
 }
 
-func startServer(bin string, port int, args []string) {
+func startServer(bin string, port int, args []string, ch chan error) {
 
 	setStatus(BuildStatus)
 	//use goroutine
 	err := build.Run(bin, args)
 	if err != nil {
-		//TODO build error
 		log.Printf("build error: ============\n%+v", err)
 		setStatus(BuildErrorStatus)
+		ch <- err
 		return
 	}
 
 	setStatus(StartupStatus)
 	//コマンドを実行
-	err = process.Run(bin)
+	err = process.Run(bin, ch)
 	if err != nil {
 		log.Printf("process error: ============\n%+v", err)
 		setStatus(StartupErrorStatus)
+		ch <- err
 		return
 	}
 
-	//TODO エラー時の挙動
 	go func() {
-		ch, err := checkConnection(port)
+		check, err := checkConnection(port)
 		if err != nil {
 			log.Printf("connection error: ============\n%+v", err)
 			setStatus(StartupErrorStatus)
 			return
 		}
 
-		v := <-ch
-		if !v {
-			setStatus(StartupErrorStatus)
-		} else {
+		v := <-check
+		if v {
+			log.Println("Complete Build and Launch")
 			setStatus(OKStatus)
+		} else {
 		}
 	}()
 }
